refactor: Files management and favicon fix
