+++ app/globals.css
@@ -50,7 +50,7 @@
 		--neutral: 0 0% 100%;
 		--foreground: 0 0% 85.49%;
 
-		--gray-100: 35 8.96 26.27%;
+		--gray-100: 35 8.96% 26.27%;
 		--gray-200: 40 12% 20%;
 		--gray-300: 38 12% 18%;
 		--gray-400: 40 8% 15%;
